refactor: main typescript
