+++ config/webpack.config.common.js
@@ -1,24 +1,9 @@
-const path = require('node:path');
 const webpack = require('webpack');
-const webpackPaths = require('./webpack.paths');
 
 /**
  * @type {webpack.Configuration}
  */
 const configuration = {
-  devtool: 'inline-source-map',
-
-  mode: 'development',
-
-  target: 'electron-renderer',
-
-  entry: [path.join(webpackPaths.srcRendererPath, 'index.tsx')],
-
-  output: {
-    path: webpackPaths.buildRendererPath,
-    filename: 'renderer.js',
-  },
-
   module: {
     rules: [
       {
