+++ package.json
@@ -2,15 +2,15 @@
   "name": "atlassify",
   "version": "1.2.0",
   "description": "Atlassian notifications on your menu bar.",
-  "main": "src/main/main.js",
+  "main": "build/main.js",
   "scripts": {
     "clean": "rimraf build coverage node_modules",
     "build": "concurrently --names 'main,renderer' --prefix-colors 'blue,green' 'pnpm build:main' 'pnpm build:renderer'",
-    "build:main": "echo 'not implemented yet'",
-    "build:renderer": "webpack --config ./config/webpack.config.prod.js",
+    "build:main": "webpack --config ./config/webpack.config.main.prod.js",
+    "build:renderer": "webpack --config ./config/webpack.config.renderer.prod.js",
     "watch": "concurrently --names 'main,renderer' --prefix-colors 'blue,green' 'pnpm watch:main' 'pnpm watch:renderer'",
-    "watch:main": "echo 'not implemented yet'",
-    "watch:renderer": "webpack --watch --config ./config/webpack.config.common.js",
+    "watch:main": "webpack --watch --config ./config/webpack.config.main.base.js",
+    "watch:renderer": "webpack --watch --config ./config/webpack.config.renderer.base.js",
     "make:linux": "electron-builder --linux",
     "make:macos": "electron-builder --mac",
     "make:win": "electron-builder --win",
@@ -157,6 +157,7 @@
     "electron": "32.1.2",
     "electron-builder": "25.0.5",
     "graphql-tag": "2.12.6",
+    "html-webpack-plugin": "5.6.0",
     "husky": "9.1.6",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
@@ -167,6 +168,7 @@
     "rimraf": "6.0.1",
     "style-loader": "4.0.0",
     "tailwindcss": "3.4.13",
+    "terser-webpack-plugin": "5.3.10",
     "ts-jest": "29.2.5",
     "ts-node": "10.9.2",
     "webpack": "5.95.0",
