+++ src/renderer/index.ejs
@@ -11,8 +11,6 @@
     <div id="atlassify"></div>
   </body>
 
-  <script src="./../../build/js/renderer.js"></script>
-
   <style>
     html,
     body,
