+++ src/renderer/index.tsx
@@ -3,7 +3,7 @@ import { createRoot } from 'react-dom/client';
 import 'tailwindcss/tailwind.css';
 import '@atlaskit/css-reset';
 
-import { App } from './app';
+import { App } from './App';
 
 const container = document.getElementById('atlassify');
 const root = createRoot(container);
